requirements.txt
fastapi
uvicorn
pydantic
python-multipart
sentence-transformers
torch
transformers
whisper
opencv-python
Pillow
clip-anytorch
streamlit
requests
